+++ tests/test-common.js
@@ -1,9 +1,4 @@
-/**
- * Test runner for JSON-LD Signatures library.
- *
- * @author Dave Longley <dlongley@digitalbazaar.com>
- * @author Manu Sporny <msporny@digitalbazaar.com>
- *
+/*!
  * Copyright (c) 2014-2018 Digital Bazaar, Inc. All rights reserved.
  */
 
@@ -23,192 +18,127 @@ module.exports = function(options) {
 
 'use strict';
 
-const assert = options.assert;
-const jsonld = options.jsonld;
-const jsigs = options.jsigs;
+const {assert, jsonld, jsigs, mock, suites, util} = options;
 const {ProofPurposeHandler} = jsigs;
-
-var testLoader = function(url, callback) {
-  if(url === testPublicKeyUrl) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKey,
-      documentUrl: testPublicKeyUrl
-    });
-  }
-  if(url === testPublicKeyUrl2) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKey2,
-      documentUrl: testPublicKeyUrl2
-    });
-  }
-  if(url === testPublicKeyOwner.id) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKeyOwner,
-      documentUrl: testPublicKeyOwner.id
-    });
-  }
-  if(url === testPublicKeyOwner2.id) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKeyOwner2,
-      documentUrl: testPublicKeyOwner2.id
-    });
-  }
-  if(url === testPublicKeyEd25519.id) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKeyEd25519,
-      documentUrl: testPublicKeyEd25519.id
-    });
-  }
-  if(url === testPublicKeyEd25519Owner.id) {
-    return callback(null, {
-      contextUrl: null,
-      document: testPublicKeyEd25519Owner,
-      documentUrl: testPublicKeyEd25519Owner.id
-    });
-  }
-};
+const {NoOpProofPurpose, NOOP_PROOF_PURPOSE_URI} = mock;
 
 // setup
-jsonld.documentLoader = testLoader;
-jsigs.use('jsonld', jsonld);
+jsonld.documentLoader = mock.testLoader;
 
 // helper:
 function clone(obj) {
   return JSON.parse(JSON.stringify(obj));
 }
 
-const noOpPpUri = 'https://example.org/special-authentication';
-
-class NoOpProofPurpose extends ProofPurposeHandler {
-  constructor(injector) {
-    super(injector);
-    // if the value of `uri` is *not* defined in SECURITY_CONTEXT then it must
-    // be in expanded form as demonstrated here
-    this.uri = noOpPpUri;
-  }
-  async validate({document, proof, purposeParameters}) {
-    return {valid: true};
-  }
-  // the proof provided here is compacted into the SECURITY_CONTEXT
-  async createProof({proof, purposeParameters}) {
-    // TODO: We may not want to mutate the proof passed in
-    proof.proofPurpose = this.uri;
-    // the proof returned here *must* be compacted into the SECURITY_CONTEXT
-    return proof;
-  }
-}
-
-jsigs.proofPurposes.use('NoOpProofPurpose', NoOpProofPurpose);
-
 // run tests
 describe('JSON-LD Signatures', function() {
-  context('common', function() {
-    const forge = jsigs.use('forge');
-
-    var testDocument = {
-      '@context': {
-        schema: 'http://schema.org/',
-        name: 'schema:name',
-        homepage: 'schema:url',
-        image: 'schema:image'
-      },
-      name: 'Manu Sporny',
-      homepage: 'https://manu.sporny.org/',
-      image: 'https://manu.sporny.org/images/manu.png'
-    };
-
-    var testBadDocument = clone(testDocument);
-    testBadDocument['https://w3id.org/security#signature'] = {
-      '@type': 'https://w3id.org/security#BogusSignature3000',
-      'http://purl.org/dc/terms/created': {
-        '@type': 'http://www.w3.org/2001/XMLSchema#dateTime',
-        '@value': '2017-03-25T22:01:04Z'
-      },
-      'http://purl.org/dc/terms/creator': {
-        '@id': 'test:1234'
-      },
-      'https://w3id.org/security#signatureValue': 'test'
-    };
-
-    it('should fail sign with unknown algorithm', function(done) {
-      jsigs.sign(testDocument, {
-        algorithm: 'BogusSignature3000',
-        privateKeyPem: '',
-        creator: ''
-      }, function(err, signedDocument) {
-        assert(err);
-        done();
-      });
-    });
-
-    it('should fail verify with unknown algorithm', function(done) {
-      jsigs.verify(testBadDocument, {}, function(err, result) {
-        assert.ifError(err);
-        assert.equal(result.verified, false, 'signature verification passed');
-        done();
-      });
-    });
-
-    it('should base64url encode', function(done) {
-      var inputs = [
-        '',
-        '1',
-        '12',
-        '123',
-        '1234',
-        '12345',
-        '\xc3\xbb\xc3\xb0\x00',
-        '\xc3\xbb\xc3\xb0',
-        '\xc3\xbb'
-      ];
-      inputs.forEach(function(input) {
-        var enc = jsigs._encodeBase64Url(input, {forge});
-        var dec = jsigs._decodeBase64Url(enc, {forge});
-        /*
-        console.log('E', input, '|', Buffer.from(input));
-        console.log('  enc', enc, '|', Buffer.from(enc));
-        console.log('  dec', dec, '|', Buffer.from(dec));
-        */
-        assert.equal(enc.indexOf('+'), -1);
-        assert.equal(enc.indexOf('/'), -1);
-        assert.equal(enc.indexOf('='), -1);
-        assert.equal(input, dec);
-      });
-      done()
-    });
-
-    it('should base64url decode', function(done) {
-      var inputs = [
-        'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_',
-        '_-E',
-        '_-E=',
-        'AA',
-        'eA',
-        'eA=',
-        'eA==',
-      ];
-      inputs.forEach(function(input) {
-        var dec = jsigs._decodeBase64Url(input, {forge});
-        var enc = jsigs._encodeBase64Url(dec, {forge});
-        /*
-        console.log('D', input, '|', Buffer.from(input));
-        console.log('  dec', dec, '|', Buffer.from(dec));
-        console.log('  enc', enc, '|', Buffer.from(enc));
-        */
-        assert.equal(input.replace(/=/g, ''), enc);
-      });
-      done()
-    });
+  context('util', function() {
+    
+    
   });
+  
+  // context('common', function() {
+  //   const forge = jsigs.use('forge');
+
+  //   const testDocument = {
+  //     '@context': {
+  //       schema: 'http://schema.org/',
+  //       name: 'schema:name',
+  //       homepage: 'schema:url',
+  //       image: 'schema:image'
+  //     },
+  //     name: 'Manu Sporny',
+  //     homepage: 'https://manu.sporny.org/',
+  //     image: 'https://manu.sporny.org/images/manu.png'
+  //   };
+
+  //   const testBadDocument = clone(testDocument);
+  //   testBadDocument['https://w3id.org/security#signature'] = {
+  //     '@type': 'https://w3id.org/security#BogusSignature3000',
+  //     'http://purl.org/dc/terms/created': {
+  //       '@type': 'http://www.w3.org/2001/XMLSchema#dateTime',
+  //       '@value': '2017-03-25T22:01:04Z'
+  //     },
+  //     'http://purl.org/dc/terms/creator': {
+  //       '@id': 'test:1234'
+  //     },
+  //     'https://w3id.org/security#signatureValue': 'test'
+  //   };
+
+  //   it('should fail sign with unknown algorithm', function(done) {
+  //     jsigs.sign(testDocument, {
+  //       algorithm: 'BogusSignature3000',
+  //       privateKeyPem: '',
+  //       creator: ''
+  //     }, function(err, signedDocument) {
+  //       assert(err);
+  //       done();
+  //     });
+  //   });
+
+  //   it('should fail verify with unknown algorithm', function(done) {
+  //     jsigs.verify(testBadDocument, {}, function(err, result) {
+  //       assert.ifError(err);
+  //       assert.equal(result.verified, false, 'signature verification passed');
+  //       done();
+  //     });
+  //   });
+
+  //   it('should base64url encode', function(done) {
+  //     const inputs = [
+  //       '',
+  //       '1',
+  //       '12',
+  //       '123',
+  //       '1234',
+  //       '12345',
+  //       '\xc3\xbb\xc3\xb0\x00',
+  //       '\xc3\xbb\xc3\xb0',
+  //       '\xc3\xbb'
+  //     ];
+  //     inputs.forEach(function(input) {
+  //       const enc = jsigs._encodeBase64Url(input, {forge});
+  //       const dec = jsigs._decodeBase64Url(enc, {forge});
+  //       /*
+  //       console.log('E', input, '|', Buffer.from(input));
+  //       console.log('  enc', enc, '|', Buffer.from(enc));
+  //       console.log('  dec', dec, '|', Buffer.from(dec));
+  //       */
+  //       assert.equal(enc.indexOf('+'), -1);
+  //       assert.equal(enc.indexOf('/'), -1);
+  //       assert.equal(enc.indexOf('='), -1);
+  //       assert.equal(input, dec);
+  //     });
+  //     done()
+  //   });
+
+  //   it('should base64url decode', function(done) {
+  //     const inputs = [
+  //       'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_',
+  //       '_-E',
+  //       '_-E=',
+  //       'AA',
+  //       'eA',
+  //       'eA=',
+  //       'eA==',
+  //     ];
+  //     inputs.forEach(function(input) {
+  //       const dec = jsigs._decodeBase64Url(input, {forge});
+  //       const enc = jsigs._encodeBase64Url(dec, {forge});
+  //       /*
+  //       console.log('D', input, '|', Buffer.from(input));
+  //       console.log('  dec', dec, '|', Buffer.from(dec));
+  //       console.log('  enc', enc, '|', Buffer.from(enc));
+  //       */
+  //       assert.equal(input.replace(/=/g, ''), enc);
+  //     });
+  //     done();
+  //   });
+  // });
 
   context('with NO security context', function() {
     // the test document that will be signed
-    var testDocument = {
+    const testDocument = {
       '@context': {
         schema: 'http://schema.org/',
         name: 'schema:name',
@@ -219,7 +149,7 @@ describe('JSON-LD Signatures', function() {
       homepage: 'https://manu.sporny.org/',
       image: 'https://manu.sporny.org/images/manu.png'
     };
-    var testDocumentSigned = {};
+    let testDocumentSigned = {};
 
     describe('signing and verify Graph2012', function() {
       it('should successfully sign a local document', function(done) {
@@ -284,7 +214,6 @@ describe('JSON-LD Signatures', function() {
             assert.ifError(err);
           }).then(done, done);
         });
-
     });
 
     describe('signing and verify LinkedDataSignature2015', function() {
@@ -376,7 +305,7 @@ describe('JSON-LD Signatures', function() {
           });
 
         it('verify local document using getPublicKey and getPublicKeyOwner ' +
-          'w/Promises API',function(done) {
+          'w/Promises API', function(done) {
           jsigs.sign(testDocument, {
             algorithm: 'LinkedDataSignature2015',
             privateKeyPem: testPrivateKeyPem3,
@@ -435,7 +364,7 @@ describe('JSON-LD Signatures', function() {
               assert.ifError(err);
               assert.equal(result.keyResults[0].error, undefined);
               assert.isBoolean(result.verified);
-              assert.isTrue(result.verified,'signature verification failed');
+              assert.isTrue(result.verified, 'signature verification failed');
               assert.isArray(result.keyResults);
               assert.equal(result.keyResults.length, 2);
               assert.isObject(result.keyResults[0]);
@@ -1078,12 +1007,12 @@ describe('JSON-LD Signatures', function() {
 
     describe('signing and verify Ed25519Signature2018', function() {
 
-      var testDocument;
-      var testDocumentSigned;
-      var testDocumentSignedAltered;
-      var testDocumentWithProofPurposeSigned;
-      var testInvalidPublicKey;
-      var testProofPurpose;
+      let testDocument;
+      let testDocumentSigned;
+      let testDocumentSignedAltered;
+      let testDocumentWithProofPurposeSigned;
+      let testInvalidPublicKey;
+      let testProofPurpose;
       let testPublicKey;
       let testPublicKeyOwner;
 
@@ -1375,17 +1304,19 @@ describe('JSON-LD Signatures', function() {
         }).then(done, done);
       });
 
-      it('should successfully sign a local document w/proofPurpose' +
+      it.only('should successfully sign a local document w/proofPurpose' +
         ' w/promises API', function(done) {
         const date = testDocumentWithProofPurposeSigned
           ['https://w3id.org/security#proof']['@graph']
           ['http://purl.org/dc/terms/created']['@value'];
         jsigs.sign(testDocument, {
-          algorithm: 'Ed25519Signature2018',
-          privateKeyBase58: testPrivateKeyEd25519Base58,
-          creator: testPublicKey.id,
-          purpose: 'NoOpProofPurpose',
-          date
+          documentLoader: testLoader,
+          suite: new suites.Ed25519Signature2018({
+            creator: testPublicKey.id,
+            date,
+            privateKeyBase58: testPrivateKeyEd25519Base58
+          }),
+          purpose: new NoOpProofPurpose()
         }).then(function(signedDocument) {
           assert.notEqual(
             signedDocument['https://w3id.org/security#proof'], undefined,