+++ tests/test-karma.js
@@ -28,8 +28,10 @@ const assert = require('chai').assert;
 const common = require('./test-common');
 const jsigs = require('..');
 const jsonld = require('../node_modules/jsonld/dist/jsonld.js');
+const mock = require('./mock-data');
+const {suites} = require('../lib/suites');
+const util = require('../lib/util');
 
-//var system = require('system');
 const forge = require('../node_modules/node-forge');
 window.forge = forge;
 const bitcoreMessage = require(
@@ -39,9 +41,12 @@ window.bitcoreMessage = bitcoreMessage;
 jsigs.promises({api: jsigs.promises});
 
 const options = {
-  assert: assert,
-  jsigs: jsigs,
-  jsonld: jsonld,
+  assert,
+  jsigs,
+  jsonld,
+  mock,
+  suites,
+  util,
   nodejs: false
 };